Windows: AFSExamineObject() refcnt underflows
Now that the reference counting is likely to be correct, do not
garbage collect objects with negative reference counts.  If the
reference counts are wrong the objects will never be destroyed but
that is now a safer choice than freeing memory that might be in use.

Reviewed-on: http://gerrit.openafs.org/9670
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
Tested-by: Jeffrey Altman <jaltman@your-file-system.com>
(cherry picked from commit 36fee95e32835cf20286c5225e56b303dd4563da)

Change-Id: Idafa8a58f362cfb082ddaee32715b1ff8550249d
Reviewed-on: http://gerrit.openafs.org/9730
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>